Add in sanitization
