+++ sdk/core/core-http/lib/policies/logPolicy.ts
@@ -3,7 +3,6 @@
 
 import { HttpOperationResponse } from "../httpOperationResponse";
 import { WebResource } from "../webResource";
-import { URLBuilder, URLQuery } from "../url";
 import {
   BaseRequestPolicy,
   RequestPolicy,
@@ -12,6 +11,7 @@ import {
 } from "./requestPolicy";
 import { Debugger } from "@azure/logger";
 import { logger as coreLogger } from "../log";
+import { Sanitizer } from "../util/sanitizer";
 
 export interface LogPolicyOptions {
   /**
@@ -34,37 +34,6 @@ export interface LogPolicyOptions {
   logger?: Debugger;
 }
 
-const RedactedString = "REDACTED";
-
-const defaultAllowedHeaderNames = [
-  "x-ms-client-request-id",
-  "x-ms-return-client-request-id",
-  "x-ms-useragent",
-  "traceparent",
-
-  "Accept",
-  "Cache-Control",
-  "Connection",
-  "Content-Length",
-  "Content-Type",
-  "Date",
-  "ETag",
-  "Expires",
-  "If-Match",
-  "If-Modified-Since",
-  "If-None-Match",
-  "If-Unmodified-Since",
-  "Last-Modified",
-  "Pragma",
-  "Request-Id",
-  "Retry-After",
-  "Server",
-  "Transfer-Encoding",
-  "User-Agent"
-];
-
-const defaultAllowedQueryParameters: string[] = ["api-version"];
-
 export function logPolicy(loggingOptions: LogPolicyOptions = {}): RequestPolicyFactory {
   return {
     create: (nextPolicy: RequestPolicy, options: RequestPolicyOptions) => {
@@ -75,9 +44,23 @@ export function logPolicy(loggingOptions: LogPolicyOptions = {}): RequestPolicyF
 
 export class LogPolicy extends BaseRequestPolicy {
   logger: Debugger;
+  sanitizer: Sanitizer;
 
-  public allowedHeaderNames: Set<string>;
-  public allowedQueryParameters: Set<string>;
+  public get allowedHeaderNames(): Set<string> {
+    return this.sanitizer.allowedHeaderNames;
+  }
+
+  public set allowedHeaderNames(allowedHeaderNames: Set<string>) {
+    this.sanitizer.allowedHeaderNames = allowedHeaderNames;
+  }
+
+  public get allowedQueryParameters(): Set<string> {
+    return this.sanitizer.allowedQueryParameters;
+  }
+
+  public set allowedQueryParameters(allowedQueryParameters: Set<string>) {
+    this.sanitizer.allowedQueryParameters = allowedQueryParameters;
+  }
 
   constructor(
     nextPolicy: RequestPolicy,
@@ -90,17 +73,7 @@ export class LogPolicy extends BaseRequestPolicy {
   ) {
     super(nextPolicy, options);
     this.logger = logger;
-
-    allowedHeaderNames = Array.isArray(allowedHeaderNames)
-      ? defaultAllowedHeaderNames.concat(allowedHeaderNames)
-      : defaultAllowedHeaderNames;
-
-    allowedQueryParameters = Array.isArray(allowedQueryParameters)
-      ? defaultAllowedQueryParameters.concat(allowedQueryParameters)
-      : defaultAllowedQueryParameters;
-
-    this.allowedHeaderNames = new Set(allowedHeaderNames.map(n => n.toLowerCase()));
-    this.allowedQueryParameters = new Set(allowedQueryParameters.map(p => p.toLowerCase()));
+    this.sanitizer = new Sanitizer({ allowedHeaderNames, allowedQueryParameters });
   }
 
   public sendRequest(request: WebResource): Promise<HttpOperationResponse> {
@@ -111,87 +84,12 @@ export class LogPolicy extends BaseRequestPolicy {
   }
 
   private logRequest(request: WebResource) {
-    this.logger(`Request: ${JSON.stringify(request, this.sanitize.bind(this), 2)}`);
-  }
-
-  private sanitize(key: string, value: unknown) {
-    if (key === "_headersMap") {
-      return this.sanitizeHeaders(key, value as {});
-    } else if (key === "url") {
-      return this.sanitizeUrl(value as string);
-    } else if (key === "query") {
-      return this.sanitizeQuery(value as {});
-    } else if (key === "body") {
-      // Don't log the request body
-      return undefined;
-    } else if (key === "response") {
-      // Don't log response again
-      return undefined;
-    } else if (key === "operationSpec") {
-      // When using sendOperationRequest, the request carries a massive
-      // field with the autorest spec. No need to log it.
-      return undefined;
-    }
-
-    return value;
-  }
-
-  private sanitizeHeaders(_: string, value: { [s: string]: any }) {
-    return this.sanitizeObject(value, this.allowedHeaderNames, (v, k) => v[k].value);
-  }
-
-  private sanitizeQuery(value: { [s: string]: string }) {
-    return this.sanitizeObject(value, this.allowedQueryParameters, (v, k) => v[k]);
-  }
-
-  private sanitizeObject(
-    value: { [s: string]: any },
-    allowedKeys: Set<string>,
-    accessor: (value: any, key: string) => any
-  ) {
-    if (typeof value !== "object" || value === null) {
-      return value;
-    }
-
-    const sanitized: { [s: string]: string } = {};
-
-    for (const k of Object.keys(value)) {
-      if (allowedKeys.has(k.toLowerCase())) {
-        sanitized[k] = accessor(value, k);
-      } else {
-        sanitized[k] = RedactedString;
-      }
-    }
-
-    return sanitized;
-  }
-
-  private sanitizeUrl(value: string): string {
-    if (typeof value !== "string" || value === null) {
-      return value;
-    }
-
-    const urlBuilder = URLBuilder.parse(value);
-    const queryString = urlBuilder.getQuery();
-
-    if (!queryString) {
-      return value;
-    }
-
-    const query = URLQuery.parse(queryString);
-    for (const k of query.keys()) {
-      if (!this.allowedQueryParameters.has(k.toLowerCase())) {
-        query.set(k, RedactedString);
-      }
-    }
-
-    urlBuilder.setQuery(query.toString());
-    return urlBuilder.toString();
+    this.logger(`Request: ${this.sanitizer.sanitize(request)}`);
   }
 
   private logResponse(response: HttpOperationResponse): HttpOperationResponse {
     this.logger(`Response status code: ${response.status}`);
-    this.logger(`Headers: ${JSON.stringify(response.headers, this.sanitize.bind(this), 2)}`);
+    this.logger(`Headers: ${this.sanitizer.sanitize(response.headers)}`);
     return response;
   }
 }
