+++ sdk/core/core-http/lib/util/inspect.browser.ts
@@ -1,0 +1,4 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export const custom = {};
