+++ sdk/core/core-http/lib/util/inspect.ts
@@ -1,0 +1,6 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import { inspect } from "util";
+
+export const custom = inspect.custom;
